Client/requirements.txt
Flask
python-dotenv
requests
PyYAML
